refactor: Move integration tests back to src
